+++ mock-engine-store/src/mock_cluster/cluster_ext.rs
@@ -8,7 +8,8 @@ use engine_store_ffi::ffi::{
     interfaces_ffi::{EngineStoreServerHelper, RaftProxyStatus, RaftStoreProxyFFIHelper},
     RaftStoreProxyFFI,
 };
-use engine_traits::Engines;
+use engine_tiflash::DB;
+use engine_traits::{Engines, KvEngine};
 use raftstore::store::RaftRouter;
 use tikv::config::TikvConfig;
 use tikv_util::{debug, sys::SysQuota, HandyRwLock};
@@ -254,6 +255,31 @@ impl<T: Simulator<TiFlashEngine>> Cluster<T> {
             // Always at the front of the vector since iterate from 0.
             self.register_ffi_helper_set(Some(0), *node_id);
         }
+    }
+}
+
+// TiFlash specific
+impl<T: Simulator<TiFlashEngine>> Cluster<T> {
+    pub fn run_conf_change_no_start(&mut self) -> u64 {
+        self.create_engines();
+        self.bootstrap_conf_change()
+    }
+
+    pub fn set_expected_safe_ts(&mut self, leader_safe_ts: u64, self_safe_ts: u64) {
+        self.cluster_ext.test_data.expected_leader_safe_ts = leader_safe_ts;
+        self.cluster_ext.test_data.expected_self_safe_ts = self_safe_ts;
+    }
+
+    pub fn get_tiflash_engine(&self, node_id: u64) -> &TiFlashEngine {
+        &self.engines[&node_id].kv
+    }
+
+    pub fn get_engines(&self, node_id: u64) -> &Engines<TiFlashEngine, engine_rocks::RocksEngine> {
+        &self.engines[&node_id]
+    }
+
+    pub fn get_raw_engine(&self, node_id: u64) -> Arc<DB> {
+        Arc::clone(self.engines[&node_id].kv.bad_downcast())
     }
 }
